+++ checkers/inc/PlayerE.h
@@ -2,7 +2,7 @@
 
 # include <cstdint>
 
-enum class PlayerE: uint8_t
+enum class PlayerE: int32_t
 {
   White = 0,
   Black = 1
